# openapi/src/main/resources/static/schema/api.yaml
openapi: "3.0.4"
info:
  title: Klavers Jassen
  description: Klavers Jassen
  version: "@git.commit.id.abbrev@"
servers:
  - url: 'https://api.klaversjassen.nl/api/v1'
    description: Klavers Jassen
  - url: 'https://api.klaversjassen.nl'
    description: Klavers Jassen Root
paths:
  /feedback:
    parameters:
      - $ref: "#/components/parameters/AppIdentifier"
    post:
      operationId: feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
              required:
                - message
      responses:
        "200":
          description: ok
  /boom:
    parameters:
      - $ref: "#/components/parameters/AppIdentifier"
    get:
      operationId: getBooms
      responses:
        "200":
          $ref: "#/components/responses/BoomsResponse"
    post:
      operationId: createBoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBoom"
      responses:
        "200":
          description: Boomresponse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boom"
  /boom/{boomId}/play:
    parameters:
      - $ref: "#/components/parameters/AppIdentifier"
    post:
      operationId: playBoom
      parameters:
        - $ref: "#/components/parameters/boomId"
      responses:
        "200":
          description: a boom
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /boom/{boomId}:
    parameters:
      - $ref: "#/components/parameters/AppIdentifier"
    get:
      operationId: getBoom
      parameters:
        - $ref: "#/components/parameters/boomId"
      responses:
        "200":
          description: a boom
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boom"
        '404':
          description: Todo item not found.
    delete:
      operationId: deleteBoom
      parameters:
        - $ref: "#/components/parameters/boomId"
      responses:
        "200":
          description: ok
...
  /games/{gameId}/claimRoem:
    parameters:
      - $ref: "#/components/parameters/AppIdentifier"
    post:
      operationId: claimRoem
      parameters:
        - $ref: "#/components/parameters/gameId"
      responses:
        "200":
          description: ok
        "422":
          description: some application error
        "404":
          description: not found
  /games/{gameId}/claimVerzaken:
    parameters:
      - $ref: "#/components/parameters/AppIdentifier"
    post:
      operationId: claimVerzaken
      parameters:
        - $ref: "#/components/parameters/gameId"
      responses:
        "200":
          description: ok
        "422":
          description: some application error
        "404":
          description: not found
  /games/{gameId}/openLastTrick:
    parameters:
      - $ref: "#/components/parameters/AppIdentifier"
    post:
      operationId: openLastTrick
      parameters:
        - $ref: "#/components/parameters/gameId"
      responses:
        "200":
          description: ok
        "422":
          description: some application error
        "404":
          description: not found
  /games/{gameId}/closeLastTrick:
    parameters:
      - $ref: "#/components/parameters/AppIdentifier"
    post:
      operationId: closeLastTrick
      parameters:
        - $ref: "#/components/parameters/gameId"
      responses:
        "200":
          description: ok
        "422":
          description: some application error
        "404":
          description: not found
  /games/{gameId}/playCard:
    parameters:
      - $ref: "#/components/parameters/AppIdentifier"
    post:
      operationId: playCard
      parameters:
        - $ref: "#/components/parameters/gameId"
      requestBody:
        $ref: "#/components/requestBodies/playCard"
      responses:
        "200":
          description: ok
        "422":
          description: some application error
        "404":
          description: not found
...
  /games:
    parameters:
      - $ref: "#/components/parameters/AppIdentifier"
    get:
      operationId: getGames
      parameters:
        - name: boom
          in: query
          schema:
            type: boolean
        - name: finished
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          $ref: "#/components/responses/GamesResponse"
    post:
      operationId: createGame
      requestBody:
        description: A new game
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createGame"
      responses:
        "200":
          $ref: "#/components/responses/GameResponse"
        "404":
          description: not found
  /games/{gameId}:
    parameters:
      - $ref: "#/components/parameters/AppIdentifier"
    get:
      operationId: getGame
      parameters:
        - $ref: "#/components/parameters/gameId"
      responses:
        "200":
          $ref: "#/components/responses/GameResponse"
        "404":
          description: not found
    delete:
      operationId: deleteGame
      parameters:
        - $ref: "#/components/parameters/gameId"
      responses:
        "200":
          description: ok
...
components:
  requestBodies:
...
    playCard:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/playCard"
    #    pingPongRequestBody:
    #      required: true
    #      content:
    #        application/json:
    #          schema:
    #            type: object
    #            properties:
    #              appIdentifier:
    #                type: string
    #            required:
    #              - appIdentifier
...
  parameters:
    AppIdentifier:
      name: App-Identifier
      in: header
      schema:
        type: string
      required: true
    gameId:
      name: gameId
      in: path
      schema:
        type: string
      required: true
...
    boomId:
      name: boomId
      in: path
      schema:
        type: string
      required: true
...
  securitySchemes:
    openId:
      type: openIdConnect
      scheme: bearer
      bearerFormat: JWT
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration
  responses:
    GameResponse:
      description: a game
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Game"
...
    BoomsResponse:
      description: all booms
      content:
        application/json:
          schema:
            type: object
            properties:
              booms:
                type: array
                items:
                  type: string
    GamesResponse:
      description: all games
      content:
        application/json:
          schema:
            type: object
            properties:
              games:
                type: array
                items:
                  type: string
...
  schemas:
    Boom:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the test
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        creator:
          description: initial dealer
          type: string
        dealer:
          type: integer
          minimum: 0
          maximum: 3
        players:
          type: array
          minItems: 4
          maxItems: 4
          items:
            type: string
        games:
          type: array
          maxItems: 32
          items:
            type: string
        gameVariant:
          $ref: '#/components/schemas/GameVariant'
        aiRisc:
          $ref: "#/components/schemas/AiRisc"
      required:
        - id
        - created
        - updated
        - creator
        - dealer
        - players
        - games
        - gameVariant
        - aiRisc
    CreateBoom:
      type: object
      properties:
        players:
          type: array
          minItems: 4
          maxItems: 4
          items:
            type: string
      required:
        - players
...
    playCard:
      type: object
      properties:
        card:
          $ref: "#/components/schemas/Card"
      required:
        - card
...
    createGame:
      type: object
      properties:
        players:
          type: array
          minItems: 4
          maxItems: 4
          items:
            type: string
      required:
        - players
...
    GameVariant:
      enum:
        - amsterdams
        - rotterdams
        - haags
...
    AiRisc:
      enum:
        - verylow
        - low
        - medium
        - high
        - veryhigh
    Game:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the test
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        creator:
          type: string
        dealer:
          type: integer
          minimum: 0
          maximum: 3
        playerCard:
          type: array
          items:
            type: object
            properties:
              card:
                $ref: '#/components/schemas/Card'
              player:
                type: integer
                minimum: 0
                maximum: 3
            required:
              - card
              - player
        #        elder:
        #          type: integer
        #          minimum: 0
        #          maximum: 3
        says:
          type: object
          additionalProperties:
            type: boolean
        players:
          type: array
          minItems: 0
          maxItems: 4
          items:
            type: string
        turns:
          type: array
          items:
            $ref: "#/components/schemas/Card"
          minItems: 0
          maxItems: 32
        trump:
          $ref: '#/components/schemas/Suit'
        whosTurn:
          type: integer
        whoSay:
          type: integer
        lastTrickOpen:
          type: boolean
        variant:
          $ref: '#/components/schemas/GameVariant'
        dealCounter:
          type: integer
        numberOfTurns:
          type: integer
        isCompleted:
          type: boolean
        hasFullTrick:
          type: boolean
        tricksPlayed:
          type: integer
        iedereenPast:
          type: boolean
        niemandGezegd:
          type: boolean
        geenKaartGespeeld:
          type: boolean
        currentTrickWinnerCard:
          $ref: "#/components/schemas/Card"
        currentTrickCards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        trickPoints:
          type: array
          items:
            $ref: "#/components/schemas/NorthSouthNumber"
        allPoints:
          $ref: "#/components/schemas/NorthSouthNumber"
        totalPoints:
          $ref: "#/components/schemas/NorthSouthNumber"
        trickRoem:
          type: array
          items:
            $ref: "#/components/schemas/NorthSouthNumber"
        allRoem:
          $ref: "#/components/schemas/NorthSouthNumber"
        totalString:
          $ref: "#/components/schemas/NorthSouthString"
        winner:
          $ref: "#/components/schemas/Teams"
        erIsGegaan:
          type: boolean
        elder:
          type: integer
        elderTeam:
          $ref: "#/components/schemas/Teams"
        trickWinnerPlayer:
          type: array
          items:
            type: integer
        trickWinnerCard:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        playerPoints:
          type: array
          items:
            type: integer
        heeftGepast:
          type: array
          items:
            type: integer
        roemGeklopt:
          type: array
          items:
            type: integer
          uniqueItems: true
        pit:
          type: boolean
        boomId:
          type: string
        verzaakt:
          type: array
          items:
            type: array
            items:
              type: integer
            uniqueItems: true
        tricksWonNorthSouth:
          type: integer
        tricksWonEastWest:
          type: integer
        aiRisc:
          $ref: "#/components/schemas/AiRisc"
      required:
        - id
        - trump
        - created
        - updated
        - creator
        - dealer
        - playerCard
        - ended
        - players
        - turns
        - trump
        - lastTrickOpen
        - variant
        - dealCounter
        - numberOfTurns
        - isCompleted
        - hasFullTrick
        - tricksPlayed
        - iedereenPast
        - niemandGezegd
        - geenKaartGespeeld
        - currentTrickCards
        - trickPoints
        - allPoints
        - trickRoem
        - allRoem
        - erIsGegaan
        - trickWinnerPlayer
        - trickWinnerCard
        - playerPoints
        - heeftGepast
        - roemGeklopt
        - verzaakt
        - aiRisc
        - totalString
    Suit:
      type: string
      enum:
        - Clubs
        - Diamonds
        - Hearts
        - Spades
    Teams:
      type: string
      enum:
        - NorthSouth
        - EastWest
    NorthSouthNumber:
      type: object
      properties:
        northSouth:
          type: integer
          default: 0
        eastWest:
          type: integer
          default: 0
      required:
        - northSouth
        - eastWest
    NorthSouthString:
      type: object
      properties:
        northSouth:
          type: string
        eastWest:
          type: string
      required:
        - northSouth
        - eastWest
    Card:
      type: string
      enum:
        - Ah
        - Kh
        - Qh
        - Jh
        - Th
        - Nh
        - Eh
        - Sh
        - Ac
        - Kc
        - Qc
        - Jc
        - Tc
        - Nc
        - Ec
        - Sc
        - Ad
        - Kd
        - Qd
        - Jd
        - Td
        - Nd
        - Ed
        - Sd
        - As
        - Ks
        - Qs
        - Js
        - Ts
        - Ns
        - Es
        - Ss
...
security:
  - bearerAuth:
      - openid
      - profile
      - email